Isolando metodo New de computers em Repository
